Project description
Write a library of static methods that perform various geometric transforms on polygons. Mathematically, a polygon is defined by its sequence of vertices (x0, y 0), (x 1, y 1), (x 2, y 2), …. In Java, we will represent a polygon by storing the x– and y-coordinates of the vertices in two parallel arrays x[] and y[]. Note that the transformation methods scale(), translate() and rotate() mutate the polygons.