Project Description
On RecursiveAppend.java write a recursive method appendNTimes that receives two arguments, a string and an integer. The method appendNTimes returns the original string appended to the original string n times. The following restrictions apply to method appendNTimes:
YOUR CODE MUST BE RECURSIVE
Do not use loops (while, do/while, or for).
Your code must return a string without extra space, comma or any other character that is not in the original string